+++ CLAUDE.md
@@ -116,7 +116,7 @@ Auth is handled by `better-auth` embedded in the API server. Supported providers
 - Real-time log streaming via SSE from the API
 - No direct database access
 - Styling: **Tailwind CSS v4** (CSS-first config via `@import "tailwindcss"` in `globals.css`) + **Radix UI primitives** (components live in `src/components/ui/`) + **Geist** font (via the `geist` npm package)
-- **Always use Radix UI primitives** for interactive components where a primitive exists (dropdown menus, dialogs, collapsibles, checkboxes, separators, etc.). Do not re-implement these with manual state + DOM event listeners — Radix handles accessibility, keyboard navigation, and focus management. Installed primitives: `@radix-ui/react-dialog`, `@radix-ui/react-separator`, `@radix-ui/react-slot`, `@radix-ui/react-dropdown-menu`, `@radix-ui/react-collapsible`, `@radix-ui/react-checkbox`, `@radix-ui/react-select`. UI components: `select.tsx`, `textarea.tsx` (matches `Input` styling, use for multi-line fields). When adding new interactive components, check for a Radix primitive first and install it if available.
+- **Always use Radix UI primitives** for interactive components where a primitive exists (dropdown menus, dialogs, collapsibles, checkboxes, separators, etc.). Do not re-implement these with manual state + DOM event listeners — Radix handles accessibility, keyboard navigation, and focus management. Installed primitives: `@radix-ui/react-dialog`, `@radix-ui/react-separator`, `@radix-ui/react-slot`, `@radix-ui/react-dropdown-menu`, `@radix-ui/react-collapsible`, `@radix-ui/react-checkbox`, `@radix-ui/react-select`, `@radix-ui/react-tabs`. UI components: `select.tsx`, `textarea.tsx` (matches `Input` styling, use for multi-line fields), `tabs.tsx` (use for tabbed views — never use native `<select>` or `<input type="radio">` for tab-like switching), `form-error.tsx` (use for inline form/action error messages — icon + subtle red background). When adding new interactive components, check for a Radix primitive first and install it if available.
 
 ### `apps/controller` (Go)
 - Watches the database for pending deployments (polling or notify)
