+++ apps/ui/src/components/ui/tabs.tsx
@@ -1,0 +1,8 @@
+"use client"
+
+import * as TabsPrimitive from "@radix-ui/react-tabs"
+
+export const Tabs = TabsPrimitive.Root
+export const TabsList = TabsPrimitive.List
+export const TabsTrigger = TabsPrimitive.Trigger
+export const TabsContent = TabsPrimitive.Content
